fixed some testcases

+++ src/instruction.rs
@@ -261,7 +261,7 @@ impl_SliceIterator_wrapper!(
 
 impl<'a> InsnDetail<'a> {
     /// Returns the implicit read registers
-    pub fn regs_read(&self) -> RegsIter<u8> {
+    pub fn regs_read(&self) -> RegsIter<u16> {
         RegsIter((*self.0).regs_read[..self.regs_read_count() as usize].iter())
     }
 
@@ -271,7 +271,7 @@ impl<'a> InsnDetail<'a> {
     }
 
     /// Returns the implicit write registers
-    pub fn regs_write(&self) -> RegsIter<u8> {
+    pub fn regs_write(&self) -> RegsIter<u16> {
         RegsIter((*self.0).regs_write[..self.regs_write_count() as usize].iter())
     }
